requirements.txt
asgiref==3.9.2
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
djangorestframework==3.16.1
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pypdfium2==4.30.0
pytesseract==0.3.13
python-decouple==3.8
python-magic-bin==0.4.14
sqlparse==0.5.3
tzdata==2025.2
